requirements.txt
aliyun-python-sdk-core
asgiref
attrs
autobahn
Automat
certifi
cffi
charset-normalizer
constantly
coreapi
coreschema
cryptography
daphne
Django
django-comment-migrate
django-cors-headers
django-filter
django-ranged-response
django-redis
django-simple-captcha
djangorestframework
djangorestframework-simplejwt
drf-yasg
hyperlink
idna
incremental
inflection
itypes
Jinja2
jmespath
MarkupSafe
packaging
Pillow
pyasn1
pyasn1-modules
pycparser
pycryptodome
PyJWT
#PyMySQL
mysqlclient
pyOpenSSL
pyparsing
pytz
redis
requests
service-identity
six
sqlparse
txaio
typing-extensions
ua-parser
uritemplate
urllib3
user-agents
xlrd
xlwt
zope.interface
pycryptodomex
jpush
tencentcloud-sdk-python
psutil
chardet
channels
channels_redis
paramiko
celery
django-celery-beat
django-celery-results
eventlet
django-simpleui